update typo, general tidy up
