+++ test/e2e/page-objects/pages/home/asset-list.ts
@@ -204,21 +204,18 @@ class AssetListPage {
     });
   }
 
-   /**
+  /**
    * This function checks if the specified token is displayed in the token list by its name.
    *
    * @param tokenName - The name of the token to check for.
    * @returns A promise that resolves if the specified token is displayed.
    */
-   async check_tokenIsDisplayed(tokenName: string): Promise<void> {
+  async check_tokenIsDisplayed(tokenName: string): Promise<void> {
     console.log(`Waiting for token ${tokenName} to be displayed`);
-    await this.driver.wait(async () => {
-      const tokenIsDisplayed = await this.driver.findElement({
-        text: tokenName,
-        tag: 'p',
-      });
-      return tokenIsDisplayed !== null;
-    }, 10000);
+    const tokenIsDisplayed = await this.driver.waitForSelector({
+      text: tokenName,
+      tag: 'p',
+    });
     console.log(`Token ${tokenName} is displayed.`);
   }
